refactoring the hooks
