+++ examples/todomvc/package.json
@@ -10,7 +10,9 @@
     "preview": "vite preview"
   },
   "dependencies": {
+    "@fireproof/core": "workspace:^",
     "@w3ui/react-keyring": "^3.0.1",
+    "@web3-storage/capabilities": "^3.2.0",
     "@web3-storage/upload-client": "^6.0.0",
     "async": "^3.2.4",
     "classnames": "^2.3.2",
@@ -22,8 +24,7 @@
     "react-router": "^6.8.1",
     "react-router-dom": "^6.8.1",
     "todomvc-app-css": "^2.4.2",
-    "todomvc-common": "^1.0.5",
-    "@fireproof/core": "workspace:^"
+    "todomvc-common": "^1.0.5"
   },
   "devDependencies": {
     "@types/react": "^18.0.27",
