+++ examples/todomvc/src/components/AllLists.tsx
@@ -2,7 +2,8 @@ import React from 'react'
 import { useContext } from 'react'
 import { Link, useLoaderData } from 'react-router-dom'
 import InputArea from './InputArea'
-import { FireproofCtx, useRevalidatorAndSubscriber, TimeTravel } from '../hooks/useFireproof'
+import { FireproofCtx, useRevalidatorAndSubscriber } from '../hooks/useFireproof'
+import { TimeTravel } from './TimeTravel'
 import { UploadManager, UploaderCtx } from '../hooks/useUploader'
 import { useKeyring } from '@w3ui/react-keyring'
 import { ListDoc } from '../interfaces'
