+++ examples/todomvc/src/components/List.tsx
@@ -5,7 +5,8 @@ import { useNavigate, useParams, useLoaderData } from 'react-router-dom'
 import Footer from './Footer'
 import InputArea from './InputArea'
 import TodoItem from './TodoItem'
-import { FireproofCtx, useRevalidatorAndSubscriber, TimeTravel } from '../hooks/useFireproof'
+import { FireproofCtx, useRevalidatorAndSubscriber } from '../hooks/useFireproof'
+import { TimeTravel } from './TimeTravel'
 import { UploadManager, UploaderCtx } from '../hooks/useUploader'
 
 import { ListLoaderData, TodoDoc } from '../interfaces'
